Create subslices for each sound
This is the 1st part of my mini series relating to Detecting Respiratory Disease with the use of Respiratory Audio (breathing sounds). For this kernel, we're only going to slice each audio file into subslices which is defined by the txt files.

    Part 2: Split into train and test
    Part 3: Create spectrogram images from audio
    Part 4: Create and train a VGG16 model with the spec images

Let's start.

Here you'll see that i'm importing librosa and soundfile which are python packages that deals with audio files.

    Librosa
    Soundfile

Boxplot_stats will be used later to see outliers. More of this later